requirements.txt
pydantic
snowflake-connector-python
pytest
fastapi
uvicorn
pyyaml
dotenv



